reside getBytes method into the JmsMapMessageUtils
